refactor(sprites): definition file in yaml
