+++ client/src/segments/__mocks__/sprite_defs.json
